Remove Jupter notebook-related code
- Tests of old cleaning scripts in Jupyter notebooks.
- Remove jupyter and nbformat requirements from setup.cfg.
- Remove .historical.util.run_notebook.
- Remove .historical.scripts.util.managers; move ColumnName enum to
  .historical.util temporarily.
