+++ item/historical/__init__.py
@@ -8,7 +8,7 @@ import pycountry
 import yaml
 
 from item.common import paths
-from item.historical.scripts.util.managers.dataframe import ColumnName
+from item.historical.util import ColumnName
 from item.remote import OpenKAPSARC, get_sdmx
 
 log = logging.getLogger(__name__)
