+++ setup.cfg
@@ -47,8 +47,6 @@ doc =
 eppa =
     gdx >= 3
 hist =
-    jupyter
-    nbformat
     requests
     requests-cache
 tests =
